CompasToolkit.jl

CompasToolkit.CompasToolkit — Module

This module provides a Julia interface to the COMPAS C++ library for simulating MRI sequences.

CompasToolkit.CartesianTrajectory — Type

CartesianTrajectory <: Trajectory

A Cartesian trajectory.

CompasToolkit.CompasArray — Type

CompasArray{T, N} <: AbstractArray{T, N}

Object representing an array of dimensionality N and type T.

CompasToolkit.CompasArray — Method

CompasArray(input::Array{ComplexF32, N})::CompasArray{ComplexF32, N} where {N}

Creates a new COMPAS array from a Julia array of ComplexF32.

CompasToolkit.CompasArray — Method

CompasArray(input::Array{Float32, N})::CompasArray{Float32, N} where {N}

Creates a new COMPAS array from a Julia array of Float32.

CompasToolkit.Context — Type

Context

A context object that manages the COMPAS C++ library.

CompasToolkit.FispSequence — Type

FispSequence

A FISP sequence.

CompasToolkit.SpiralTrajectory — Type

SpiralTrajectory <: Trajectory

A Spiral trajectory.

CompasToolkit.TissueParameters — Type

TissueParameters

Tissue parameters.

CompasToolkit.Trajectory — Type

Abstract type for trajectories.

CompasToolkit.pSSFPSequence — Type

pSSFPSequence

A pSSFP sequence.

Base.collect — Method

Base.collect(input::CompasArray{ComplexF32, N}) where {N}

Collects the data from a COMPAS array of ComplexF32 into a Julia array.

Base.collect — Method

Base.collect(input::CompasArray{Float32, N}) where {N}

Collects the data from a COMPAS array of Float32 into a Julia array.

CompasToolkit.assert_size — Method

assert_size(input::AbstractArray, expected::Dims)

Asserts that the size of the input array matches the expected size. Throws an error if the sizes do not match.

CompasToolkit.check_version — Method

check_version()

Checks if the version of the COMPAS C++ library matches the expected version. Throws an error if the versions do not match.

CompasToolkit.compute_jacobian — Method

compute_jacobian(
    echos::AbstractMatrix,
    𝜕echos::NamedTuple{(:T1, :T2)},
    parameters::TissueParameters,
    trajectory::Trajectory,
    coils::AbstractMatrix,
    v::AbstractMatrix
)::CompasArray{ComplexF32, 2}

Computes the Hermitian Jacobian of the magnetization with respect to the tissue parameters, using a FISP sequence.

CompasToolkit.compute_jacobian_hermitian — Method

compute_jacobian_hermitian(
    echos::AbstractMatrix,
    𝜕echos::NamedTuple{(:T1, :T2)},
    parameters::TissueParameters,
    trajectory::Trajectory,
    coils::AbstractMatrix,
    v::AbstractArray{<:Any,3}
)::CompasArray{ComplexF32, 2}

Computes the Hermitian Jacobian of the magnetization with respect to the tissue parameters, using a pSSFP sequence.

CompasToolkit.convert_array — Method

convert_array(input::Array{T,N}, ty::Type{T}, dims::Integer...)::CompasArray{T,N} where {T,N}

Converts a Julia array to a COMPAS array of the specified type and dimensions.

CompasToolkit.convert_array — Method

convert_array(input::CompasArray{T,N}, ty::Type{T}, dims::Integer...)::CompasArray{T,N} where {T,N}

Converts a COMPAS array to a COMPAS array of the specified type and dimensions.

CompasToolkit.convert_array — Method

convert_array(input::AbstractArray, ty::Type{T}, dims::Integer...,)::CompasArray where {T}

Converts an abstract array to a COMPAS array of the specified type and dimensions.

CompasToolkit.convert_array_host — Method

convert_array_host(
    input::AbstractArray,
    ty::Type{T},
    dims::Dims{N},
)::Array{T,N} where {T,N}

Converts an abstract array to a Julia array of the specified type and dimensions.

CompasToolkit.convert_array_host — Method

convert_array_host(
    input::Array{T,N},
    ty::Type{T},
    dims::Dims{N},
)::Array{T,N} where {T,N}

Converts a Julia array to a Julia array of the specified type and dimensions.

CompasToolkit.convert_array_host — Method

convert_array_host(input::Number, ty::Type{T}, dims::Dims{N})::Array{T,N} where {T,N}

Converts a number to a Julia array of the specified type and dimensions.

CompasToolkit.get_context — Method

get_context()::Context

Gets the current COMPAS context. Throws an error if the COMPAS toolkit has not been initialized.

CompasToolkit.init_context — Function

init_context(device::Integer)::Context

Initializes a COMPAS context on the specified device and sets it as the current context.

CompasToolkit.magnetization_to_signal — Method

magnetization_to_signal(
    echos::AbstractMatrix,
    parameters::TissueParameters,
    trajectory::CartesianTrajectory,
    coils::AbstractMatrix,
)::CompasArray{ComplexF32, 3}

Converts the magnetization of a tissue sample to a signal.

CompasToolkit.magnetization_to_signal — Method

magnetization_to_signal(
    echos::AbstractMatrix,
    parameters::TissueParameters,
    trajectory::SpiralTrajectory,
    coils::AbstractMatrix,
)::CompasArray{ComplexF32, 3}

Converts the magnetization of a tissue sample to a signal.

CompasToolkit.set_context — Method

set_context(context::Context)

Sets the COMPAS context for the current thread.

CompasToolkit.set_context — Method

set_context(context::Context)

Sets the COMPAS context for the current thread.

CompasToolkit.simulate_magnetization — Method

simulate_magnetization(
    parameters::TissueParameters,
    sequence::FispSequence,
)::CompasArray{ComplexF32, 2}

Simulates the magnetization of a tissue sample using a FISP sequence.

CompasToolkit.simulate_magnetization — Method

simulate_magnetization(
    parameters::TissueParameters,
    sequence::pSSFPSequence,
)::CompasArray{ComplexF32, 2}

Simulates the magnetization of a tissue sample using a pSSFP sequence.

CompasToolkit.simulate_magnetization_derivative — Method

simulate_magnetization_derivative(
    field::Integer,
    echos::AbstractMatrix{ComplexF32},
    parameters::TissueParameters,
    sequence::FispSequence,
    Δ::AbstractFloat
)::CompasArray{ComplexF32, 2}

Simulates the derivative of the magnetization of a tissue sample with respect to a field, using a FISP sequence.

CompasToolkit.simulate_magnetization_derivative — Method

simulate_magnetization_derivative(
    field::Integer,
    echos::AbstractMatrix{ComplexF32},
    parameters::TissueParameters,
    sequence::pSSFPSequence,
    Δ::AbstractFloat
)::CompasArray{ComplexF32, 2}

Simulates the derivative of the magnetization of a tissue sample with respect to a field, using a pSSFP sequence.

CompasToolkit.simulate_magnetization_derivatives — Method

simulate_magnetization_derivatives(
    echos::AbstractMatrix{ComplexF32},
    parameters::TissueParameters,
    sequence,
    Δ::AbstractFloat
)::@NamedTuple{
    T1::CompasArray{ComplexF32, 2},
    T2::CompasArray{ComplexF32, 2}
}

Simulates the derivatives of the magnetization of a tissue sample with respect to T1 and T2.

CompasToolkit.simulate_magnetization_derivatives — Method

simulate_magnetization_derivatives(
    echos::AbstractMatrix{ComplexF32},
    parameters::TissueParameters,
    sequence,
)::@NamedTuple{
    T1::CompasArray{ComplexF32, 2},
    T2::CompasArray{ComplexF32, 2}
}

Simulates the derivatives of the magnetization of a tissue sample with respect to T1 and T2, using a default Δ value of 1e-4.

CompasToolkit.synchronize — Method

synchronize()

Waits until all asynchronous operations of Compas have finished.

CompasToolkit.unsafe_destroy_object! — Method

unsafe_destroy_object!(obj)

Destroys a COMPAS object. This function is unsafe because it does not check if the object is valid.

CompasToolkit.version — Method

version()::String

Returns the version of the COMPAS C++ library.